Struct amethyst_utils::circular_buffer::CircularBuffer

pub struct CircularBuffer<A> { /* fields omitted */ }

A CircularBuffer that drops the oldest element inserted when full.

Methods

impl<A> CircularBuffer<A>

pub fn new<B>(size: usize) -> CircularBuffer<B>

Creates a new CircularBuffer with fixed size

pub fn push(&mut self, elem: A) -> Option<A>

Add a value to the CircularBuffer Returns the popped value if the buffer is full

pub fn queue(&self) -> &VecDeque<A>

Get an immutable reference to the values inside the CircularBuffer