Enum amethyst_utils::ortho_camera::CameraNormalizeMode

pub enum CameraNormalizeMode {
    Lossy {
        stretch_direction: Axis2,
    },
    Contain,
}

Settings that decide how to scale the camera's matrix when the aspect ratio changes.

Variants
Lossy

Using an aspect ratio of 1:1, tries to adjust the matrix values of the camera so that the direction opposite to the stretch_direction always have a world size of 1.

This means that the direction opposite to stretch_direction will always be between 0.0 to 1.0 in world coordinates. Scene space can be lost on the specified stretch_direction however.

Example: If you use Lossy with the stretch_direction of Axis::X, this means that a mesh or image going from the world coordinates (0, 0) to (1, 1) would take the whole screen size if the window dimension width is equal to its height.

If the window gets stretched on the x axis, the mesh or image will stay centered and the background clear color (or things that were previously outside of the window) will now be shown on the left and right sides of the mesh or image.

If you shrink the window on the x axis instead, the left and right parts of the images will go off screen and will NOT be shown.

If you want the whole world space between (0, 0) and (1, 1) to be shown at ALL times, consider using CameraNormalizeMode::Contain instead.

Fields of Lossy

stretch_direction: Axis2

Contain

Scales the render dynamically to ensure no space is lost in the [0,1] range on any axis. In other words, this ensures that you can always at least see everything between the world coordinates (0, 0) up to (1, 1).

If you have a non-default Transform on your camera, it will just translate those coordinates by the translation of the Transform.