Module amethyst_utils::removal

Provides utilities to remove large amounts of entities with a single command.

Structs

Removal
A marker Component used to remove entities and clean up your scene. The generic parameter I is the type of id you want to use. Generally an int or an enum.

Functions

add_removal_to_entity
Adds a Removal component with the specified id to the specified entity. Usually used with prefabs, when you want to add a Removal component at the root of the loaded prefab.

exec_removal
Removes all entities that have the Removal<I> component with the specified removal_id.