Function amethyst_utils::removal::add_removal_to_entity

pub fn add_removal_to_entity<T: PartialEq + Clone + Debug + Send + Sync + 'static>(
    entity: Entity, 
    id: T, 
    storage: &mut WriteStorage<'_, Removal<T>>
)

Adds a Removal component with the specified id to the specified entity. Usually used with prefabs, when you want to add a Removal component at the root of the loaded prefab.