Struct amethyst_utils::scene::BasicScenePrefab

pub struct BasicScenePrefab<V, M = ObjFormat> where
    M: Format<Mesh>,
    M::Options: DeserializeOwned + Serialize,  { /* fields omitted */ }

Basic Prefab scene node, meant to be used for fast prototyping, and most likely replaced for more complex scenarios.

Type parameters:

V: Vertex format to use for generated Meshes, must to be one of: * Vec<PosTex> * Vec<PosNormTex> * Vec<PosNormTangTex> * ComboMeshCreator

M: Format to use for loading Meshes from file