Module amethyst_utils::ortho_camera

Provides a automatically resized orthographic camera.

Structs

CameraOrtho
Component attached to the camera’s entity that allows automatically adjusting the camera’s matrix according to preferences in the “mode” and “world_coordinates” fields. It adjusts the camera so that the camera’s world coordinates are always visible. You must add the CameraOrthoSystem to your dispatcher for this to take effect (no dependencies required).

CameraOrthoSystem
System that automatically changes the camera matrix according to the settings in the CameraOrtho attached to the camera entity.

CameraOrthoWorldCoordinates
The coordinates that CameraOrtho will keep visible in the window. bottom can be a higher value than top, as is common in 2D coordinates

Enums

CameraNormalizeMode
Settings that decide how to scale the camera’s matrix when the aspect ratio changes.